app/requirements.txt
selenium
discord.py
Pillow